+++ lib/screens/channel/channel.dart
@@ -4,12 +4,11 @@ import 'package:floating/floating.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_mobx/flutter_mobx.dart';
-import 'package:frosty/api/bttv_api.dart';
-import 'package:frosty/api/ffz_api.dart';
-import 'package:frosty/api/seventv_api.dart';
-import 'package:frosty/api/twitch_api.dart';
-import 'package:frosty/constants/constants.dart';
-import 'package:frosty/core/auth/auth_store.dart';
+import 'package:frosty/apis/bttv_api.dart';
+import 'package:frosty/apis/ffz_api.dart';
+import 'package:frosty/apis/seventv_api.dart';
+import 'package:frosty/apis/twitch_api.dart';
+import 'package:frosty/constants.dart';
 import 'package:frosty/main.dart';
 import 'package:frosty/screens/channel/chat/chat.dart';
 import 'package:frosty/screens/channel/chat/details/chat_details_store.dart';
@@ -19,7 +18,8 @@ import 'package:frosty/screens/channel/video/video.dart';
 import 'package:frosty/screens/channel/video/video_overlay.dart';
 import 'package:frosty/screens/channel/video/video_store.dart';
 import 'package:frosty/screens/settings/settings.dart';
-import 'package:frosty/screens/settings/settings_store.dart';
+import 'package:frosty/screens/settings/stores/auth_store.dart';
+import 'package:frosty/screens/settings/stores/settings_store.dart';
 import 'package:frosty/widgets/modal.dart';
 import 'package:provider/provider.dart';
