+++ lib/screens/home/top/categories/categories_store.dart
@@ -1,8 +1,8 @@
 import 'dart:io';
 
-import 'package:frosty/api/twitch_api.dart';
-import 'package:frosty/core/auth/auth_store.dart';
+import 'package:frosty/apis/twitch_api.dart';
 import 'package:frosty/models/category.dart';
+import 'package:frosty/screens/settings/stores/auth_store.dart';
 import 'package:mobx/mobx.dart';
 
 part 'categories_store.g.dart';
