+++ lib/screens/settings/stores/auth_store.dart
@@ -2,10 +2,10 @@ import 'dart:async';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:frosty/api/twitch_api.dart';
-import 'package:frosty/constants/constants.dart';
-import 'package:frosty/core/user/user_store.dart';
+import 'package:frosty/apis/twitch_api.dart';
+import 'package:frosty/constants.dart';
 import 'package:frosty/main.dart';
+import 'package:frosty/screens/settings/stores/user_store.dart';
 import 'package:mobx/mobx.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:webview_flutter/webview_flutter.dart';
