+++ lib/screens/settings/stores/user_store.dart
@@ -1,4 +1,4 @@
-import 'package:frosty/api/twitch_api.dart';
+import 'package:frosty/apis/twitch_api.dart';
 import 'package:frosty/models/user.dart';
 import 'package:mobx/mobx.dart';
